pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "manim-gen"
version = "0.1.0"
description = "Natural language to Manim math animation generator"
requires-python = ">=3.11"
dependencies = [
    "manim>=0.18.0",
    "rich>=13.7",
    "httpx>=0.27",
    "pydantic>=2.6",
    "python-dotenv>=1.0",
]

[project.scripts]
manim-gen = "manim_gen.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["src/manim_gen"]
